refactor(foundation): move event definition back to plugin foundation
